refactor: move esbuild test related scripts to `scripts/src/esbuild-tests` (#7377)
Moved to `scripts/src/esbuild-tests` to share some code between the `gen-esbuild-test` script.
