+++ scripts/package.json
@@ -5,8 +5,8 @@
   "scripts": {
     "type-check": "tsc -b",
     "setup-bench": "oxnode ./misc/setup-benchmark-input/index.js",
-    "gen-esbuild-test": "oxnode ./misc/gen-esbuild-test.ts",
-    "esbuild-snap-diff": "oxnode ./snap-diff/index.ts"
+    "gen-esbuild-tests": "oxnode ./src/esbuild-tests/gen-tests.ts",
+    "esbuild-snap-diff": "oxnode ./src/esbuild-tests/snap-diff/index.ts"
   },
   "dependencies": {
     "@pnpm/find-workspace-packages": "catalog:",
