+++ scripts/src/esbuild-tests/snap-diff/aggregate-reason.ts
