+++ scripts/src/esbuild-tests/snap-diff/esbuild-snapshots/snapshots_default.txt
