+++ scripts/src/esbuild-tests/snap-diff/esbuild-snapshots/snapshots_importstar_ts.txt
