+++ scripts/src/esbuild-tests/snap-diff/esbuild-snapshots/snapshots_lower.txt
