+++ scripts/src/esbuild-tests/snap-diff/index.ts
