+++ scripts/src/esbuild-tests/snap-diff/rewrite.ts
