+++ scripts/src/esbuild-tests/snap-diff/snap-parser.ts
