+++ scripts/src/esbuild-tests/snap-diff/stats/aggregated-reason.md
