+++ scripts/src/esbuild-tests/snap-diff/summary/packagejson.md
