+++ scripts/src/esbuild-tests/snap-diff/types.ts
